bitbake: server/process: Remove daemonic thread usage
We're seeing UI deadlocks occasionally and this is possibly due to the
use of a daemonic thread in the UI event queue processing. This thread
could terminate holding a threading Lock() which would cause issues
for the process when exitting.

Change the shutdown process to handle this more cleanly.

(Bitbake rev: d65926ca381e43aeb299173007241200e1935990)

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>